Code updating by fixing tslinit error
