+++ package.json
@@ -1,46 +1,54 @@
 {
-  "name": "node-web-app",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build:backend": "tsc",
-    "build:frontend": "webpack",
-    "start": "set debug=* && ts-node-dev --respawn --transpileOnly ./src/main.ts"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/dotenv": "^6.1.1",
-    "@types/express": "^4.16.1",
-    "@types/node": "^11.13.18",
-    "@types/react": "^16.8.6",
-    "@types/react-dom": "^16.8.2",
-    "css-loader": "^2.1.0",
-    "dotenv": "^8.0.0",
-    "nodemon": "^1.19.1",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^0.23.1",
-    "ts-loader": "^5.3.3",
-    "typescript": "^3.3.3333",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.2.3"
-  },
-  "dependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "class-validator": "^0.9.1",
-    "ejs": "^2.6.1",
-    "express": "^4.16.4",
-    "jsonwebtoken": "^8.5.1",
-    "mysql": "^2.17.1",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
-    "reflect-metadata": "^0.1.13",
-    "ts-node": "^8.3.0",
-    "ts-node-dev": "^1.0.0-pre.40",
-    "typeorm": "^0.2.18"
-  }
+    "name": "node-web-app",
+    "version": "1.0.0",
+    "description": "",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "build:backend": "tsc",
+        "build:frontend": "webpack",
+        "build:backend:production": "npm run build:backend && npm prune --production",
+        "lint": "tslint  -e \"./node_modules/**/*.ts\" ./**/*.ts -c ./tslint.json",
+        "start": "ts-node-dev --respawn --transpileOnly ./src/main.ts",
+        "migration:run": "ts-node ./node_modules/typeorm/cli.js migration:run",
+        "migration:revert": "ts-node ./node_modules/typeorm/cli.js migration:revert"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "devDependencies": {
+        "@types/dotenv": "^6.1.1",
+        "@types/express": "^4.16.1",
+        "@types/node": "^11.13.18",
+        "@types/react": "^16.8.6",
+        "@types/react-dom": "^16.8.2",
+        "css-loader": "^2.1.0",
+        "dotenv": "^8.0.0",
+        "nodemon": "^1.19.1",
+        "source-map-loader": "^0.2.4",
+        "style-loader": "^0.23.1",
+        "ts-loader": "^5.3.3",
+        "tslint-auto-fix": "^1.0.0",
+        "tslint-eslint-rules": "^5.4.0",
+        "typescript": "^3.5.3",
+        "webpack": "^4.29.6",
+        "webpack-cli": "^3.2.3"
+    },
+    "dependencies": {
+        "@types/bcryptjs": "^2.4.2",
+        "backend-tslint-rules": "^1.14.0",
+        "bcryptjs": "^2.4.3",
+        "body-parser": "^1.19.0",
+        "class-validator": "^0.9.1",
+        "ejs": "^2.6.1",
+        "express": "^4.16.4",
+        "jsonwebtoken": "^8.5.1",
+        "mysql": "^2.17.1",
+        "react": "^16.8.3",
+        "react-dom": "^16.8.3",
+        "reflect-metadata": "^0.1.13",
+        "ts-node": "^8.3.0",
+        "ts-node-dev": "^1.0.0-pre.40",
+        "tslint": "^5.18.0",
+        "typeorm": "^0.2.18"
+    }
 }
