+++ src/web/controllers/home-controller.ts
@@ -1,0 +1,12 @@
+import { Request, Response } from 'express';
+import { Locals } from '../providers/Locals';
+
+class Home {
+    static index = (request: Request, response: Response): any => {
+        response.json({
+            message: Locals.config().name
+        });
+    };
+}
+
+export { Home };
