+++ src/web/entity/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "compilerOptions" {
+        "strictPropertyInitialization": false,
+        "experimentalDecorators": true
+    }
+}
