+++ src/web/entity/user.ts
@@ -1,0 +1,45 @@
+import * as bcrypt from 'bcryptjs';
+import { IsNotEmpty, Length } from 'class-validator';
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  PrimaryGeneratedColumn,
+  Unique,
+  UpdateDateColumn
+} from 'typeorm';
+
+@Entity()
+@Unique(['username'])
+export class User {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  @Length(4, 20)
+  username: string;
+
+  @Column()
+  @Length(4, 100)
+  password: string;
+
+  @Column()
+  @IsNotEmpty()
+  role: string;
+
+  @Column()
+  @CreateDateColumn()
+  createdAt: Date;
+
+  @Column()
+  @UpdateDateColumn()
+  updatedAt: Date;
+
+  hashPassword(): void {
+    this.password = bcrypt.hashSync(this.password, 8);
+  }
+
+  checkIfUnencryptedPasswordIsValid(unencryptedPassword: string): boolean {
+    return bcrypt.compareSync(unencryptedPassword, this.password);
+  }
+}
